+++ Profile/Shopware/Converter/ShippingMethodConverter.php
@@ -7,6 +7,7 @@
 
 namespace SwagMigrationAssistant\Profile\Shopware\Converter;
 
+use Shopware\Core\Checkout\Cart\Tax\TaxCalculator;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\Log\Package;
 use Shopware\Core\Framework\Rule\Container\AndRule;
@@ -14,9 +15,12 @@ use Shopware\Core\Framework\Rule\Container\OrRule;
 use SwagMigrationAssistant\Migration\Converter\ConvertStruct;
 use SwagMigrationAssistant\Migration\DataSelection\DefaultEntities;
 use SwagMigrationAssistant\Migration\Logging\Log\EmptyNecessaryFieldRunLog;
+use SwagMigrationAssistant\Migration\Logging\LoggingServiceInterface;
+use SwagMigrationAssistant\Migration\Mapping\MappingServiceInterface;
 use SwagMigrationAssistant\Migration\MigrationContextInterface;
 use SwagMigrationAssistant\Profile\Shopware\Logging\Log\UnsupportedShippingCalculationType;
 use SwagMigrationAssistant\Profile\Shopware\Logging\Log\UnsupportedShippingPriceLog;
+use SwagMigrationAssistant\Profile\Shopware\Premapping\DefaultShippingAvailabilityRule;
 
 #[Package('services-settings')]
 abstract class ShippingMethodConverter extends ShopwareConverter
@@ -61,6 +65,17 @@ abstract class ShippingMethodConverter extends ShopwareConverter
         'availabilityRuleId' => 'availability_rule_id',
     ];
 
+    private TaxCalculator $taxCalculator;
+
+    public function __construct(
+        MappingServiceInterface $mappingService,
+        LoggingServiceInterface $loggingService,
+        TaxCalculator $taxCalculator
+    ) {
+        parent::__construct($mappingService, $loggingService);
+        $this->taxCalculator = $taxCalculator;
+    }
+
     public function convert(array $data, Context $context, MigrationContextInterface $migrationContext): ConvertStruct
     {
         $this->generateChecksum($data);
@@ -73,19 +88,6 @@ abstract class ShippingMethodConverter extends ShopwareConverter
         $this->connectionId = '';
         if ($connection !== null) {
             $this->connectionId = $connection->getId();
-        }
-
-        if (!isset($data['calculation'])
-            || !\array_key_exists($data['calculation'], self::CALCULATION_TYPE_MAPPING)
-        ) {
-            $this->loggingService->addLogEntry(new UnsupportedShippingCalculationType(
-                $this->runId,
-                DefaultEntities::SHIPPING_METHOD,
-                $this->oldShippingMethod,
-                $data['calculation']
-            ));
-
-            return new ConvertStruct(null, $data);
         }
 
         $converted = [];
@@ -110,9 +112,15 @@ abstract class ShippingMethodConverter extends ShopwareConverter
             $this->mappingIds[] = $defaultDeliveryTimeMapping['id'];
         }
 
-        $defaultAvailabilityRuleUuid = $this->mappingService->getDefaultAvailabilityRule($this->context);
+        $defaultAvailabilityRuleUuid = $this->mappingService->getMapping(
+            $this->connectionId,
+            DefaultShippingAvailabilityRule::getMappingName(),
+            'default_shipping_availability_rule',
+            $this->context
+        );
         if ($defaultAvailabilityRuleUuid !== null) {
-            $converted['availabilityRuleId'] = $defaultAvailabilityRuleUuid;
+            $converted['availabilityRuleId'] = $defaultAvailabilityRuleUuid['entityUuid'];
+            $this->mappingIds[] = $defaultAvailabilityRuleUuid['id'];
         }
 
         $fields = $this->checkForEmptyRequiredConvertedFields($converted, $this->requiredDataFields);
@@ -149,8 +157,19 @@ abstract class ShippingMethodConverter extends ShopwareConverter
         }
 
         if (isset($data['shippingCosts'])) {
-            $calculationType = self::CALCULATION_TYPE_MAPPING[$data['calculation']];
-            $converted['prices'] = $this->getShippingCosts($data['shippingCosts'], $calculationType, $priceRule);
+            if (!isset($data['calculation'])
+                || !\array_key_exists($data['calculation'], self::CALCULATION_TYPE_MAPPING)
+            ) {
+                $this->loggingService->addLogEntry(new UnsupportedShippingCalculationType(
+                    $this->runId,
+                    DefaultEntities::SHIPPING_METHOD,
+                    $this->oldShippingMethod,
+                    $data['calculation']
+                ));
+            } else {
+                $calculationType = self::CALCULATION_TYPE_MAPPING[$data['calculation']];
+                $converted['prices'] = $this->getShippingCosts($data, $calculationType, $priceRule);
+            }
         }
 
         $this->setCustomAvailabilityRule($data, $converted);
@@ -555,8 +574,14 @@ abstract class ShippingMethodConverter extends ShopwareConverter
         return $rule;
     }
 
-    protected function getShippingCosts(array $shippingCosts, int $calculationType, ?array $rule): array
+    protected function getShippingCosts(array $data, int $calculationType, ?array $rule): array
     {
+        $shippingCosts = $data['shippingCosts'];
+        $taxRate = 0.0;
+        if (isset($data['tax']['tax'])) {
+            $taxRate = (float) $data['tax']['tax'];
+        }
+
         $convertedCosts = [];
         foreach ($shippingCosts as $key => $shippingCost) {
             $key = (int) $key;
@@ -612,7 +637,14 @@ abstract class ShippingMethodConverter extends ShopwareConverter
             }
 
             $this->convertValue($cost, 'quantityStart', $shippingCost, 'from', self::TYPE_FLOAT);
-            $this->convertValue($cost, 'price', $shippingCost, 'value', self::TYPE_FLOAT);
+
+            $priceNet = \round(((float) $shippingCost['value'] * 100) / (100 + $taxRate), $this->context->getRounding()->getDecimals());
+            $cost['currencyPrice'] = [[
+                'currencyId' => $currencyUuid,
+                'gross' => (float) $shippingCost['value'],
+                'net' => $priceNet,
+                'linked' => true,
+            ]];
 
             if ($rule !== null) {
                 $cost['rule'] = $rule;